Fix Phase 1 implementation based on feedback
- Require absolute paths in repository configuration schema
- Add repository identifier documentation to schema (explains URL mapping)
- Modify existing github_mcp_server.py instead of creating new file
- Add validation for absolute paths in RepositoryConfig
- Remove github_mcp_server_multi_repo.py (merged into main server)
- Update repository manager to enforce absolute path requirement
- Add comprehensive test for relative path validation
- Clean up imports and maintain backward compatibility

Fixes:
✅ Repository schema requires absolute paths with pattern validation
✅ Schema documents how repository identifiers become URL components
✅ Server changes are reviewable as diff of existing file
✅ Proper validation prevents relative path configuration errors
✅ All tests pass including new absolute path validation test
