+++ config/repositories_schema.json
@@ -1,19 +1,21 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
   "title": "GitHub Agent Repository Configuration",
-  "description": "Configuration schema for multiple repository support in GitHub MCP Agent",
+  "description": "Configuration schema for multiple repository support in GitHub MCP Agent. Repository identifiers (keys) become part of MCP server URLs: http://localhost:8080/mcp/{identifier}/",
   "type": "object",
   "properties": {
     "repositories": {
       "type": "object",
-      "description": "Dictionary of repository configurations",
+      "description": "Dictionary of repository configurations. Keys are repository identifiers used in URLs.",
       "patternProperties": {
         "^[a-zA-Z0-9_-]+$": {
           "type": "object",
+          "description": "Repository configuration. The key becomes the repository identifier in MCP URLs.",
           "properties": {
             "path": {
               "type": "string",
-              "description": "Absolute or relative path to the git repository"
+              "description": "Absolute path to the git repository (relative paths not allowed)",
+              "pattern": "^/"
             },
             "description": {
               "type": "string",
